Ehm, of course I had to forget something. I placed all the blocks left, regardless if their region was left or right, what I did was this: theme_blocks("all", $this); in my left <TD>.
Now I changed it to
theme_block("left", $this);
in the left <TD> and
theme_blocks("right", $this);
in the right. If everything was this easy.

OK, really, this theme is NS4.7 NS4.76 mozilla (but mozilla 0.8 cuz he's buggy as hell with the BACKGROUNDS in TDs) and IE comliant or compatible, whatever it's being called.

Jeroen. -- Happy.