Hi all,
I tweaked my theme a bit, don't know if it'll help much but I was able to squeeze a table out of abstract and article. So what you'd say? Well, since the abstract function is called for every post on the main page, I thought that it'd improve it a bit. Of course it isn't much but less tables to show the same is always a good thing. Also, due of the deleted table, I could delete some backgrounds (less traffic from the server to the client) and placed them once in a TABLE header. OK, but a cool side-effect is that it's more HTML 4.01 compliant. Because BACKGROUNDs are allowed in TABLEs and I think it's a bit unclear if they can be in TD yes or no. Either way, this brought to my attention that my !DOCTYPE is strict... I started of strict though... But I think it's a better idea to switch to transitional/loose. OK I think that was it... Oh also because of the changed TABLE structure I had to delete some BGCOLORs. Again, it might be a teeny weeny bit faster, but when you see my site being build up it isn't that nice anymore. (bgcolors are displayed faster than backgrounds). Nevertheless its the result that counts not? Let's hope everything still works.

Jeroen.